feat: remove kafka flush
