+++ blockchainetl/jobs/exporters/multi_item_exporter.py
@@ -19,11 +19,6 @@
 # LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
 # OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
 # SOFTWARE.
-import logging
-from collections import defaultdict
-
-from ethereumetl.enumeration.entity_type import EntityType
-
 
 class MultiItemExporter:
     def __init__(self, item_exporters):
@@ -33,31 +28,7 @@ class MultiItemExporter:
         for exporter in self.item_exporters:
             exporter.open()
 
-    @staticmethod
-    def statistics_by_block(dict_list, item_type):
-        if item_type == 'block':
-            return
-
-        count_dict = defaultdict(int)
-
-        # 遍历字典列表
-        for d in dict_list:
-          if item_type == d.get('type'):
-              count_dict[d['block_number']] += 1
-
-        for group, count in count_dict.items():
-          logging.info(f"Exporting {count} {item_type} from block {group}.")
-
     def export_items(self, items):
-        logging.info(f"-------- Exports Statistics --------")
-        blocks = set([
-          item.get('number') for item in items if item.get('type') == 'block'
-        ])
-        logging.info(f"Exporting {len(blocks)} block, {blocks}.")
-
-        for entity_type in tuple(EntityType.ALL_FOR_STREAMING):
-            self.statistics_by_block(items, entity_type)
-
         for exporter in self.item_exporters:
             exporter.export_items(items)
